+++ lib/restio.dart
@@ -11,6 +11,7 @@ export 'package:restio/src/core/auth/basic_authenticator.dart';
 export 'package:restio/src/core/auth/bearer_authenticator.dart';
 export 'package:restio/src/core/auth/digest_authenticator.dart';
 export 'package:restio/src/core/auth/hawk_authenticator.dart';
+export 'package:restio/src/core/auth/multi_authenticator.dart';
 export 'package:restio/src/core/auth/nonce.dart';
 export 'package:restio/src/core/cache/cache.dart' show Cache;
 export 'package:restio/src/core/cache/cache_store.dart';